# Improve perfromance for loading from Stream on Windows