--- a/src/System.Drawing.Common/src/Configurations.props
+++ b/src/System.Drawing.Common/src/Configurations.props
@@ -4,8 +4,6 @@
     <PackageConfigurations>
       netcoreapp2.0-Windows_NT;
       netcoreapp2.0-Unix;
-      netcoreapp2.1-Windows_NT;
-      netcoreapp2.1-Unix;
       net461;
       netstandard;
     </PackageConfigurations>