… src/System.Drawing.Common/src/System/Drawing/GdiplusNative.Windows.cs
… src/System.Drawing.Common/src/System/Drawing/Icon.Windows.cs
… src/System.Drawing.Common/src/System/Drawing/Image.Windows.cs
… src/System.Drawing.Common/src/System/Drawing/Image.cs
… src/System.Drawing.Common/src/System/Drawing/Internal/GPStream.cs
… src/System.Drawing.Common/src/System/Drawing/UnsafeNativeMethods.cs
… src/System.Drawing.Common/tests/BitmapTests.cs
… src/System.Drawing.Common/tests/Performance/Perf_Image_Load.cs
… src/System.Drawing.Common/tests/Performance/System.Drawing.Common.Performance.Tests.csproj