- Make IStream blittable and move to Interop
- Change GPStream to use Span APIs
- Make STATSTG blittable and move to Interop
- Fill out corners correctly in GPStream
- Remove extra allocations opening Gifs
- Fix allocations in FrameDimensionsList and move to shared
- Move IPicture to Icon.Windows (was Windows specific and used IStream, needs cleanup and moving to Interop)
- Add perf tests for opening images