@JeremyKuhne 2018-07-18
Before:

```
  System.Drawing.Common.Performance.Tests.dll                            | Metric                                        | Unit  | Iterations |    Average | STDEV.S |        Min |        Max
  :---------------------------------------------------------------------- |:--------------------------------------------- |:-----:|:----------:| ----------:| -------:| ----------:| ----------:
   System.Drawing.Tests.Perf_Image_Load.Bitmap_FromStream                 | Duration                                      | msec  |     5      |   2104.315 |  47.190 |   2052.298 |   2170.079
   System.Drawing.Tests.Perf_Image_Load.Bitmap_FromStream                 | Allocation Size on Benchmark Execution Thread | bytes |     5      | 1.207E+009 |   0.000 | 1.207E+009 | 1.207E+009
   System.Drawing.Tests.Perf_Image_Load.Image_FromStream                  | Duration                                      | msec  |     5      |   2061.762 |  33.282 |   2039.791 |   2118.484
   System.Drawing.Tests.Perf_Image_Load.Image_FromStream                  | Allocation Size on Benchmark Execution Thread | bytes |     5      | 1.207E+009 |   0.000 | 1.207E+009 | 1.207E+009
   System.Drawing.Tests.Perf_Image_Load.Image_FromStream_NoValidation     | Duration                                      | msec  |    100     |     11.951 |   0.623 |     11.371 |     15.602
   System.Drawing.Tests.Perf_Image_Load.Image_FromStream_NoValidation     | Allocation Size on Benchmark Execution Thread | bytes |    100     | 213600.000 |   0.000 | 213600.000 | 213600.000
   System.Drawing.Tests.Perf_Image_Load.Image_FromStream_NoValidation_Gif | Duration                                      | msec  |    100     |     20.703 |   0.606 |     19.907 |     22.184
   System.Drawing.Tests.Perf_Image_Load.Image_FromStream_NoValidation_Gif | Allocation Size on Benchmark Execution Thread | bytes |    100     | 5.438E+006 |   0.000 | 5.438E+006 | 5.438E+006
```

After:

```
   System.Drawing.Common.Performance.Tests.dll                            | Metric                                        | Unit  | Iterations |   Average | STDEV.S |       Min |       Max
  :---------------------------------------------------------------------- |:--------------------------------------------- |:-----:|:----------:| ---------:| -------:| ---------:| ---------:
   System.Drawing.Tests.Perf_Image_Load.Bitmap_FromStream                 | Duration                                      | msec  |     6      |  1883.708 |  13.440 |  1870.467 |  1902.180
   System.Drawing.Tests.Perf_Image_Load.Bitmap_FromStream                 | Allocation Size on Benchmark Execution Thread | bytes |     6      | 31200.000 |   0.000 | 31200.000 | 31200.000
   System.Drawing.Tests.Perf_Image_Load.Image_FromStream                  | Duration                                      | msec  |     6      |  1879.785 |   4.414 |  1874.196 |  1885.028
   System.Drawing.Tests.Perf_Image_Load.Image_FromStream                  | Allocation Size on Benchmark Execution Thread | bytes |     6      | 31200.000 |   0.000 | 31200.000 | 31200.000
   System.Drawing.Tests.Perf_Image_Load.Image_FromStream_NoValidation     | Duration                                      | msec  |    100     |    12.245 |   0.301 |    11.442 |    13.866
   System.Drawing.Tests.Perf_Image_Load.Image_FromStream_NoValidation     | Allocation Size on Benchmark Execution Thread | bytes |    100     | 31200.000 |   0.000 | 31200.000 | 31200.000
   System.Drawing.Tests.Perf_Image_Load.Image_FromStream_NoValidation_Gif | Duration                                      | msec  |    100     |    19.631 |   3.527 |    18.657 |    54.310
   System.Drawing.Tests.Perf_Image_Load.Image_FromStream_NoValidation_Gif | Allocation Size on Benchmark Execution Thread | bytes |    100     | 10400.000 |   0.000 | 10400.000 | 10400.000
```

A key thing to note is that most of the time spent opening images is validating them. On this particular machine (which is loading from NVM/e), validating is 170 times slower. Other measures I've seen are in the 150-200 times slower. We should look at adding the option to skip and flipping the default behavior for 3.0 (with an AppContext switch to force).

Unix can be improved as well. @safern plans to investigate.