@JeremyKuhne 2018-07-18
@dotnet-bot test Windows x64 Debug Build