@safern 2018-07-18
Thanks for doing this. Love the results.

The reason is because the package needs to be netcoreapp2.0 compatible and the netstandard implementation is a not supported assembly. However I think it would be fine now that we’ve shipped the compat pack to harvest the 2.0 assets.

@weshaggard any complains with doing that?