@ericstj 2018-07-18
If you consider the netstandard2.0 surface area done then go ahead and harvest all the old implementations and refs.

If you imagine that you still might want to grow netstandard2.0 surface area (either adding stuff already in desktop or new types) then you should keep those builds around and ifdef.