@safern 2018-07-18
> If you consider the netstandard2.0 surface area done then go ahead and harvest all the old implementations and refs.
> If you imagine that you still might want to grow netstandard2.0 surface area (either adding stuff already in desktop or new types) then you should keep those builds around and ifdef.

We're going to be doing a lot of this perf work which I think it will involve using Span and those ref-like 2.1 apis, so I think it is not worth the effort of having if defs all over the place for this perf effort only to support ns2.0, if we add new types/apis I think it would be fine for people to move to vnext.