@ericstj 2018-07-18
>  if we add new types/apis I think it would be fine for people to move to vnext.

You can not make that generalization for all packages.  It's not "move to vNext" for many customers it's "target the majority of live frameworks" to "target a single netcoreapp3.0 framework".  That's not a small thing and for many customers will be a deal breaker.  You'll need to decide that based on the business case for APIs being added.  For things built on top of netstandard you need to remember that you are not only serving netcoreapp customers, but also those which want to cross target.

I will repeat: 
1. If you know of API you want to add now that needs to be available for everyone then keep this library cross-compiling.
2. If you don't know of any API that will be added then you can opportunistically harvest to save yourself some work now, you just may need to bring it back in the future.