@weshaggard 2018-07-18
I have not issues with harvesting the older assets, but please be explicit about it as @ericstj mentioned. If we add any additional APIs to the netstandard20 ref then you will have to start building the netcoreapp2.0 asset live again in order for it to be compliant with the ref.