@safern 2018-07-18
Also, I was looking to System.Drawing.Common configurations.props and I've noticed that we aren't including netcoreapp vnext in the packageconfigurations. If we're going to if def implementation I think we should include that configuration as a packageconfiguration as well so that it takes effect on the package also when using it in vnext. @weshaggard, @ericstj correct me if I'm wrong.