@safern 2018-07-18
@ericstj I wonder why the Packaging All Configurations is failing when building for netcoreapp2.1 if the references are conditioned to `TargetsNetFx != true`